pratt: reorganize
